Add a tryOAuth2() callback in the data loader iframe.
The idea is that if we aren't authorized to get some jsonp data (missing an
auth header, cookie, etc), the data provider can call this callback and
attempt to get an authorization key for the service.  authtest.json is an
example of how this works (the example doesn't actually validate the auth=
parameter, it just checks to make sure it's there, and if not, tries to get a
Google OAuth2 token for a random Google API).