Add 'intensify' option.
If you provide this and use the default table display, numeric values will
be formatted with the cell background colour varying from red (negative
numbers) through white (0) through blue (positive numbers) automatically
depending on the min/max values in the whole grid.

This is handy for things like week/hour grids: a value measured for days of
the week across one axis, and hour of the day on the other one.

Motivated by this example:
http://www.flickr.com/photos/idvsolutions/8265288798/sizes/o/in/photostream/