Fix datatable scrollbars, and handle window resizes correctly.
If you set gviz's options.height property - even, ironically, if you set it
redundantly to the height of the container, which ought to have been obvious
- then gviz will make the headers fixed at the top of the container even
when you're scrolling, which is what we want (ie. like "freeze rows" in a
spreadsheet).

We also add a listener on resize events and elegantly auto-resize the chart
or table when it happens.