Experimental chart=heatgrid implementation.
Heatgrid is a non-gviz based visualization that plots each value in a 2d
grid as one pixel/rectangle on a chart, with the intensity of each pixel set
according to the value in that cell.  You can mouseover the chart to see the
actual values.

This isn't perfect yet, but I think this is a decent first version.