heatgrid: keyboard bindings, highlight box, more viewable popover.
The popover would go off the bottom or right edges of the screen if you
moved your mouse too far in that direction; now it tries to put itself to
the left or above your mouse cursor if you're in the right or bottom half of
the screen, respectively.

There's now a red box that surrounds whatever pixel you're trying to
highlight, giving a more precise view than the mouse pointer.

You can use the up/down/left/right arrows to more precisely adjust which
pixel you want to view.