Improve extract_regexp date handling, and add quantize and yspread functions.
extract_regexp: add support for non-string columns.  Previously it would
just give an error if you tried to extract_regexp on dates, for example.
Now it works as you might expect (it renders the date into a string first).

quantize: a function to split a given column's values into buckets.  You can
give either a bucket size (in which case it rounds down (floor()) all
numbers to a multiple of that size), or a series of bucket boundaries, in
which case it uses the exact boundaries you specify.  Useful for generating
histograms.

yspread: a function to take all the numeric columns on a given row and
re-express them as a fraction of the total value of that row, between 0 and
1.  This is useful when using chart=stacked and quantize; you can get a kind
of "histogram over time" view.

Change-Id: Id1a2d6aeaf6274b3b1cc8ba47748cc8744b0c9d0