Provide a way to send GVIZ "Configuration Options" along with a transformation to afterquery.
A feature by which a user can provide configuration options along with the transformations to be transferred to the underlying gviz chart  without having to explicitly list them in the afterquery code. This way users have a tool to, for example, change the look of the chart, activate things such as logarithmic scales, invert the axes, etc.